refactor(email_api): split calendar handlers out of sync.rs (cycle28 LOC)
