+++ src/bin/email_api_dir/external_handlers/mod.rs
@@ -3,8 +3,10 @@ pub(crate) mod openapi;
 pub(crate) mod accounts;
 pub(crate) mod folders;
 pub(crate) mod sync;
+pub(crate) mod calendar;
 
 pub(crate) use openapi::*;
 pub(crate) use accounts::*;
 pub(crate) use folders::*;
 pub(crate) use sync::*;
+pub(crate) use calendar::*;
